+++ app/charts/area/areas-state.tsx
@@ -348,7 +348,7 @@ const useAreasState = (
     marginRight: right,
   });
   const margins = {
-    top: 50 + yAxisLabelMargin,
+    top: 70 + yAxisLabelMargin,
     right,
     bottom,
     left,