+++ app/pages/_preview.tsx
@@ -2,38 +2,49 @@ import { Box } from "@mui/material";
 import { useEffect } from "react";
 
 import { useLocale } from "@/locales/use-locale";
+import CONFIGURATOR_STATE_MAP from "@/test/__fixtures/config/prod/map-1.json";
+
+const loadIframe = (id: string, config: any) => {
+  const iframe = document.getElementById(id) as HTMLIFrameElement;
+  iframe.onload = async () => {
+    const iframeWindow = iframe?.contentWindow as Window | undefined;
+
+    if (iframeWindow) {
+      iframeWindow.postMessage(config, "*");
+    }
+  };
+};
 
 const Page = () => {
   const locale = useLocale();
-  useEffect(() => {
-    const iframe = document.getElementById("chart") as HTMLIFrameElement;
-    iframe.onload = async () => {
-      const iframeWindow = iframe?.contentWindow as Window | undefined;
 
-      if (iframeWindow) {
-        iframeWindow.postMessage(CONFIGURATOR_STATE, "*");
-      }
-    };
+  useEffect(() => {
+    loadIframe("chart-column", CONFIGURATOR_STATE_COLUMN);
+    loadIframe("chart-map", CONFIGURATOR_STATE_MAP);
   }, []);
 
   return (
-    <Box sx={{ display: "flex", flexDirection: "column", gap: 6, p: 6 }}>
-      <iframe
-        id="chart"
-        src={`/${locale}/preview`}
-        style={{
+    <Box
+      sx={{
+        display: "flex",
+        gap: 6,
+        p: 6,
+        "& > iframe": {
           width: 450,
           height: 750,
           border: "none",
-        }}
-      />
+        },
+      }}
+    >
+      <iframe id="chart-column" src={`/${locale}/preview`} />
+      <iframe id="chart-map" src={`/${locale}/preview`} />
     </Box>
   );
 };
 
 export default Page;
 
-const CONFIGURATOR_STATE = {
+const CONFIGURATOR_STATE_COLUMN = {
   state: "CONFIGURING_CHART",
   dataSet:
     "https://energy.ld.admin.ch/sfoe/bfe_ogd84_einmalverguetung_fuer_photovoltaikanlagen/2",