+++ CHANGELOG.md
@@ -12,6 +12,9 @@ You can also check the
 # Unreleased
 
 - Features
+  - Centered x-axis labels
+  - Added y-axis labels
+  - Implemented Content Security Policy (CSP)
   - Added a new option to the vertical axis on the line chart, allowing users to
     display a dot over the line where it intercepts the x-axis ticks