+++ CHANGELOG.md
@@ -14,11 +14,12 @@ You can also check the
 - Features
   - Centered x-axis labels
   - Added y-axis labels
-  - Implemented Content Security Policy (CSP)
   - Added a new option to the vertical axis on the line chart, allowing users to
     display a dot over the line where it intercepts the x-axis ticks
   - Downloading images of bar charts now includes the whole chart, not just the
     visible part
+- Fixes
+  - Preview via API now works correctly for map charts
 
 # [5.1.0] - 2025-01-14
 
@@ -41,8 +42,6 @@ You can also check the
     switching between chart types
   - Improved readability for dataset selection and preview
   - Sorting dataset results by score option is now correctly available to select
-  - Created with visualize.admin.ch footnote is only displayed once, not twice,
-    when downloading an image of a published chart
   - Changing dates manually using keyboard works correctly now in date picker
     inputs
   - Opening a temporal dimension with timezone in table chart configurator